requirements.txt
dotenv==0.9.9
python-telegram-bot==22.7
pydantic==2.13.3